--- pkg/kubelet/cri/remote/OWNERS
@@ -1,6 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-approvers:
-  - feiskyer
-  - sig-node-approvers
-  - sig-node-cri-approvers
