+++ pkg/kubelet/util/util_windows.go
@@ -20,95 +20,14 @@ limitations under the License.
 package util
 
 import (
-	"context"
 	"fmt"
-	"net"
-	"net/url"
 	"path/filepath"
 	"strings"
 	"syscall"
 	"time"
-
-	"github.com/Microsoft/go-winio"
 )
 
-const (
-	tcpProtocol   = "tcp"
-	npipeProtocol = "npipe"
-)
-
-// CreateListener creates a listener on the specified endpoint.
-func CreateListener(endpoint string) (net.Listener, error) {
-	protocol, addr, err := parseEndpoint(endpoint)
-	if err != nil {
-		return nil, err
-	}
-
-	switch protocol {
-	case tcpProtocol:
-		return net.Listen(tcpProtocol, addr)
-
-	case npipeProtocol:
-		return winio.ListenPipe(addr, nil)
-
-	default:
-		return nil, fmt.Errorf("only support tcp and npipe endpoint")
-	}
-}
-
-// GetAddressAndDialer returns the address parsed from the given endpoint and a context dialer.
-func GetAddressAndDialer(endpoint string) (string, func(ctx context.Context, addr string) (net.Conn, error), error) {
-	protocol, addr, err := parseEndpoint(endpoint)
-	if err != nil {
-		return "", nil, err
-	}
-
-	if protocol == tcpProtocol {
-		return addr, tcpDial, nil
-	}
-
-	if protocol == npipeProtocol {
-		return addr, npipeDial, nil
-	}
-
-	return "", nil, fmt.Errorf("only support tcp and npipe endpoint")
-}
-
-func tcpDial(ctx context.Context, addr string) (net.Conn, error) {
-	return (&net.Dialer{}).DialContext(ctx, tcpProtocol, addr)
-}
-
-func npipeDial(ctx context.Context, addr string) (net.Conn, error) {
-	return winio.DialPipeContext(ctx, addr)
-}
-
-func parseEndpoint(endpoint string) (string, string, error) {
-	// url.Parse doesn't recognize \, so replace with / first.
-	endpoint = strings.Replace(endpoint, "\\", "/", -1)
-	u, err := url.Parse(endpoint)
-	if err != nil {
-		return "", "", err
-	}
-
-	if u.Scheme == "tcp" {
-		return "tcp", u.Host, nil
-	} else if u.Scheme == "npipe" {
-		if strings.HasPrefix(u.Path, "//./pipe") {
-			return "npipe", u.Path, nil
-		}
-
-		// fallback host if not provided.
-		host := u.Host
-		if host == "" {
-			host = "."
-		}
-		return "npipe", fmt.Sprintf("//%s%s", host, u.Path), nil
-	} else if u.Scheme == "" {
-		return "", "", fmt.Errorf("Using %q as endpoint is deprecated, please consider using full url format", endpoint)
-	} else {
-		return u.Scheme, "", fmt.Errorf("protocol %q not supported", u.Scheme)
-	}
-}
+const npipeProtocol = "npipe"
 
 // LocalEndpoint returns the full path to a named pipe at the given endpoint - unlike on unix, we can't use sockets.
 func LocalEndpoint(path, file string) (string, error) {
