+++ pkg/probe/exec/exec.go
@@ -21,8 +21,8 @@ import (
 	"errors"
 
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
+	remote "k8s.io/cri-client/pkg"
 	"k8s.io/kubernetes/pkg/features"
-	"k8s.io/kubernetes/pkg/kubelet/cri/remote"
 	"k8s.io/kubernetes/pkg/kubelet/util/ioutils"
 	"k8s.io/kubernetes/pkg/probe"
