+++ staging/publishing/import-restrictions.yaml
@@ -303,6 +303,12 @@
   - k8s.io/klog
   - k8s.io/utils
 
-# TODO: enable when source code got transferred
-# - baseImportPath: "./staging/src/k8s.io/cri-client"
-#   allowedImports: []
+- baseImportPath: "./staging/src/k8s.io/cri-client"
+  allowedImports:
+  - k8s.io/apimachinery
+  - k8s.io/apiserver
+  - k8s.io/component-base
+  - k8s.io/cri-api
+  - k8s.io/cri-client
+  - k8s.io/klog/v2
+  - k8s.io/utils
