+++ staging/src/k8s.io/cri-client/pkg/fake/fake_runtime.go
@@ -24,7 +24,7 @@ import (
 	"google.golang.org/grpc"
 	kubeapi "k8s.io/cri-api/pkg/apis/runtime/v1"
 	apitest "k8s.io/cri-api/pkg/apis/testing"
-	"k8s.io/kubernetes/pkg/kubelet/util"
+	"k8s.io/cri-client/pkg/util"
 	utilexec "k8s.io/utils/exec"
 )
