+++ staging/src/k8s.io/cri-client/pkg/remote_image_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2017 The Kubernetes Authors.
+Copyright 2022 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,63 +14,42 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package remote
+package cri
 
 import (
 	"context"
 	"os"
 	"testing"
-	"time"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 	oteltrace "go.opentelemetry.io/otel/trace"
 	"go.opentelemetry.io/otel/trace/noop"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	internalapi "k8s.io/cri-api/pkg/apis"
-	apitest "k8s.io/cri-api/pkg/apis/testing"
+	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1"
+	"k8s.io/cri-client/pkg/util"
 	"k8s.io/klog/v2"
-	fakeremote "k8s.io/kubernetes/pkg/kubelet/cri/remote/fake"
-	"k8s.io/kubernetes/pkg/kubelet/util"
 )
 
-const (
-	defaultConnectionTimeout = 15 * time.Second
-)
-
-// createAndStartFakeRemoteRuntime creates and starts fakeremote.RemoteRuntime.
-// It returns the RemoteRuntime, endpoint on success.
-// Users should call fakeRuntime.Stop() to cleanup the server.
-func createAndStartFakeRemoteRuntime(t *testing.T) (*fakeremote.RemoteRuntime, string) {
-	endpoint, err := fakeremote.GenerateEndpoint()
-	require.NoError(t, err)
-
-	fakeRuntime := fakeremote.NewFakeRemoteRuntime()
-	fakeRuntime.Start(endpoint)
-
-	return fakeRuntime, endpoint
-}
-
-func createRemoteRuntimeService(endpoint string, t *testing.T) internalapi.RuntimeService {
+func createRemoteImageServiceWithTracerProvider(endpoint string, tp oteltrace.TracerProvider, t *testing.T) internalapi.ImageManagerService {
 	logger := klog.Background()
-	runtimeService, err := NewRemoteRuntimeService(endpoint, defaultConnectionTimeout, noop.NewTracerProvider(), &logger)
-
+	runtimeService, err := NewRemoteImageService(endpoint, defaultConnectionTimeout, tp, &logger)
 	require.NoError(t, err)
 
 	return runtimeService
 }
 
-func createRemoteRuntimeServiceWithTracerProvider(endpoint string, tp oteltrace.TracerProvider, t *testing.T) internalapi.RuntimeService {
+func createRemoteImageServiceWithoutTracerProvider(endpoint string, t *testing.T) internalapi.ImageManagerService {
 	logger := klog.Background()
-	runtimeService, err := NewRemoteRuntimeService(endpoint, defaultConnectionTimeout, tp, &logger)
+	runtimeService, err := NewRemoteImageService(endpoint, defaultConnectionTimeout, noop.NewTracerProvider(), &logger)
 	require.NoError(t, err)
 
 	return runtimeService
 }
 
-func TestGetSpans(t *testing.T) {
+func TestImageServiceSpansWithTP(t *testing.T) {
 	fakeRuntime, endpoint := createAndStartFakeRemoteRuntime(t)
 	defer func() {
 		fakeRuntime.Stop()
@@ -86,15 +65,17 @@ func TestGetSpans(t *testing.T) {
 		sdktrace.WithBatcher(exp),
 	)
 	ctx := context.Background()
-	rtSvc := createRemoteRuntimeServiceWithTracerProvider(endpoint, tp, t)
-	_, err := rtSvc.Version(ctx, apitest.FakeVersion)
+	imgSvc := createRemoteImageServiceWithTracerProvider(endpoint, tp, t)
+	imgRef, err := imgSvc.PullImage(ctx, &runtimeapi.ImageSpec{Image: "busybox"}, nil, nil)
+	assert.NoError(t, err)
+	assert.Equal(t, "busybox", imgRef)
 	require.NoError(t, err)
 	err = tp.ForceFlush(ctx)
 	require.NoError(t, err)
 	assert.NotEmpty(t, exp.GetSpans())
 }
 
-func TestVersion(t *testing.T) {
+func TestImageServiceSpansWithoutTP(t *testing.T) {
 	fakeRuntime, endpoint := createAndStartFakeRemoteRuntime(t)
 	defer func() {
 		fakeRuntime.Stop()
@@ -105,11 +86,17 @@ func TestVersion(t *testing.T) {
 			}
 		}
 	}()
-
+	exp := tracetest.NewInMemoryExporter()
+	tp := sdktrace.NewTracerProvider(
+		sdktrace.WithBatcher(exp),
+	)
 	ctx := context.Background()
-	rtSvc := createRemoteRuntimeService(endpoint, t)
-	version, err := rtSvc.Version(ctx, apitest.FakeVersion)
+	imgSvc := createRemoteImageServiceWithoutTracerProvider(endpoint, t)
+	imgRef, err := imgSvc.PullImage(ctx, &runtimeapi.ImageSpec{Image: "busybox"}, nil, nil)
+	assert.NoError(t, err)
+	assert.Equal(t, "busybox", imgRef)
 	require.NoError(t, err)
-	assert.Equal(t, apitest.FakeVersion, version.Version)
-	assert.Equal(t, apitest.FakeRuntimeName, version.RuntimeName)
+	err = tp.ForceFlush(ctx)
+	require.NoError(t, err)
+	assert.Empty(t, exp.GetSpans())
 }
