+++ staging/src/k8s.io/cri-client/pkg/util/util_unsupported.go
@@ -20,25 +20,17 @@ limitations under the License.
 package util
 
 import (
+	"context"
 	"fmt"
-	"time"
+	"net"
 )
 
-// LockAndCheckSubPath empty implementation
-func LockAndCheckSubPath(volumePath, subPath string) ([]uintptr, error) {
-	return []uintptr{}, nil
+// CreateListener creates a listener on the specified endpoint.
+func CreateListener(endpoint string) (net.Listener, error) {
+	return nil, fmt.Errorf("CreateListener is unsupported in this build")
 }
 
-// UnlockPath empty implementation
-func UnlockPath(fileHandles []uintptr) {
-}
-
-// LocalEndpoint empty implementation
-func LocalEndpoint(path, file string) (string, error) {
-	return "", fmt.Errorf("LocalEndpoints are unsupported in this build")
-}
-
-// GetBootTime empty implementation
-func GetBootTime() (time.Time, error) {
-	return time.Time{}, fmt.Errorf("GetBootTime is unsupported in this build")
+// GetAddressAndDialer returns the address parsed from the given endpoint and a context dialer.
+func GetAddressAndDialer(endpoint string) (string, func(ctx context.Context, addr string) (net.Conn, error), error) {
+	return "", nil, fmt.Errorf("GetAddressAndDialer is unsupported in this build")
 }
